requirements.txt
streamlit>=1.40
openai>=1.0
chromadb>=0.5
python-dotenv
